move update_args() to utils
