+++ batch_inference.py
@@ -2,7 +2,7 @@ import os
 import yaml
 import argparse
 
-from utils import is_push_to_hf_hub_enabled
+from utils import is_push_to_hf_hub_enabled, update_args
 from src.pipeline.batch_inference import gen_local_lm_responses
 from src.pipeline.utils import push_to_hf_hub
 
@@ -40,31 +40,6 @@ def batch_inference(args):
         )
 
     return local_lm_responses
-
-def update_args(parser, args):
-    """
-    update_args replaces default values of the args. Arguments that are directly
-    specified from CLI has more priority. 
-    
-    For instance, if the value of an argument has default value, and if the value of 
-    that argument is defined in YAML file, the latter value will be kept. 
-
-    If the value of an argument is specified manually/directly from the CLI (not default),
-    but if the value of that argument is defined in YAML file, the latter value will be 
-    ignored. Hence, the former value will be kept.
-    """
-    if args.from_config is not None:
-        with open(args.from_config, 'r') as file:
-            config_from_yaml = yaml.safe_load(file)
-
-        for key, value in config_from_yaml.items():
-            if hasattr(args, key):  # Check if the attribute exists
-                current_value = getattr(args, key)
-                default_value = parser.get_default(key)
-                if current_value == default_value:
-                    setattr(args, key, value)
-
-    return args
 
 if __name__ == "__main__":    
     parser = argparse.ArgumentParser(description="CLI for batch inference step")
